Remove space between labels
According to spec labels are separated by a comma, without any whitespace.
To show this behavior we changed the integration test to have one case
with 2 labels. This makes the integrationtest fail.

We are using the `withLabels` method twice because their is an incompatibility
between php 7.3 and 7.4 vardict arguments are different ordered between those versions.
By calling the method twice we can make sure the order is kept.